Complete functions for the game
(I think that) I finished the functions to play the game properly, which means my next step next week will be to implement this in pygame or with another package to make it look pretty. So please let me know if you think something is missing, or if it's incomplete!